copied linux readl/writel for accessing registers, not I can set the MAC/PHY state correctly

+++ src/llamaos/net/e1000e/CSR.h
@@ -150,7 +150,7 @@ private:
    CSR (const CSR &);
    CSR &operator= (const CSR &);
 
-   uint32_t *const pointer;
+   uint8_t *const pointer;
 
 };
